refactor: core:ui 모듈 의존성 제거
