+++ android/Bottari/core/designsystem/build.gradle.kts
@@ -7,7 +7,5 @@ android {
 }
 
 dependencies {
-    implementation(project(":core:ui"))
-
     implementation(libs.androidx.appcompat)
 }
