+++ android/Bottari/presentation/build.gradle.kts
@@ -52,7 +52,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":core:designsystem"))
     implementation(project(":core:ui"))
 
     implementation(libs.androidx.core.ktx)
