+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/home/HomeBottomNavigationBar.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.bottari.core.designsystem.theme.BottariTheme
-import com.bottari.core.ui.source.NoRippleInteractionSource
+import com.bottari.core.designsystem.interaction.NoRippleInteractionSource
 import com.bottari.presentation.R
 import com.bottari.presentation.compose.common.navigation.Screen
