+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/home/template/bookmark/BookmarkTemplateScreen.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.bottari.core.designsystem.theme.BottariTheme
-import com.bottari.core.ui.modifier.noRippleClickable
+import com.bottari.core.designsystem.modifier.noRippleClickable
 import com.bottari.core.ui.modifier.topBottomFadingEdge
 import com.bottari.presentation.compose.home.template.component.TemplateEmptyView
 import com.bottari.presentation.compose.home.template.component.TemplateItem
