+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/template/component/SelectedBottariSection.kt
@@ -20,7 +20,7 @@ import com.bottari.core.designsystem.chip.BottariChip
 import com.bottari.core.designsystem.component.BottariBox
 import com.bottari.core.designsystem.theme.BottariTheme
 import com.bottari.core.ui.modifier.dashedBorder
-import com.bottari.core.ui.modifier.noRippleClickable
+import com.bottari.core.designsystem.modifier.noRippleClickable
 
 @Composable
 fun SelectedBottariSection(
